refactor(core-weave): extract progression resolvers
Move page-definition, KnopInventory, and MeshInventory progression resolution
out of weave.ts into progression_resolvers.ts while preserving planner behavior
and public façade imports.
